@@ -361,7 +361,7 @@ class Exchanges::HbxProfilesController < ApplicationController
     @element_to_replace_id = params[:family_actions_id]
     createSep
     respond_to do |format|
-      format.js { render :file => "sep/approval/add_sep_result.js.erb", name: @name }
+      format.js { render "sep/approval/add_sep_result.js.erb", name: @name }
     end
   end
 
@@ -380,7 +380,7 @@ class Exchanges::HbxProfilesController < ApplicationController
     @family_id = params_parser.family_id
     params_parser.cancel_enrollments
     respond_to do |format|
-      format.js { render :file => "datatables/cancel_enrollment_result.js.erb"}
+      format.js { render "datatables/cancel_enrollment_result.js.erb"}
     end
   end
 
@@ -399,7 +399,7 @@ class Exchanges::HbxProfilesController < ApplicationController
     @family_id = params_parser.family_id
     params_parser.terminate_enrollments
     respond_to do |format|
-      format.js { render :file => "datatables/terminate_enrollment_result.js.erb"}
+      format.js { render "datatables/terminate_enrollment_result.js.erb"}
     end
   end
 
