@@ -5,7 +5,7 @@ require "rails_helper"
 RSpec.describe "insured/plan_shoppings/plans.js.erb" do
   before :each do
     assign :plans, []
-    render :file => "insured/plan_shoppings/plans.js.erb"
+    render template: "insured/plan_shoppings/plans.js.erb"
   end
 
   it "should call aptc" do
