@@ -1,6 +1,6 @@
 class UserMailer < ApplicationMailer
-  ### add_template_helper makes the view helper methods available in the Mailer templates. It does NOT make the methods available in the Mailer itself
-  ### Thus we have to use Include in addition to add_template_helper
+  ### helper makes the view helper methods available in the Mailer templates. It does NOT make the methods available in the Mailer itself
+  ### Thus we have to use Include in addition to helper
   helper Config::AcaHelper
   helper Config::SiteHelper
   helper Config::ContactCenterHelper
