@@ -76,7 +76,7 @@ RSpec.describe "insured/plan_shoppings/receipt.html.erb" do
     @plan = plan_cost_decorator
     allow(@plan).to receive(:sole_source?).and_return(true)
     allow(view).to receive(:policy_helper).and_return(double('FamilyPolicy', updateable?: true))
-    render file: "insured/plan_shoppings/receipt.en.html.erb"
+    render template: "insured/plan_shoppings/receipt.en.html.erb"
   end
 
   it "should match the data on the confirmation receipt" do
