@@ -13,7 +13,7 @@ RSpec.describe "products/qhp/summary.js.erb" do
     assign(:hbx_enrollment, hbx_enrollment)
     allow(benefit_group).to receive(:sole_source?).and_return(true)
     stub_template "shared/_summary.html.erb" => ''
-    render file: "products/qhp/summary.js.erb"
+    render template: "products/qhp/summary.js.erb"
   end
 
   it "should call account-detail" do
